@@ -11,6 +11,7 @@ import RevenueCat
 import Purchases
 import Persistence
 import MediaManager
+import AppIntents
 import OSLog
 
 @main
@@ -45,20 +46,37 @@ struct ShotbotApp: App {
             guard phase == .active else { return }
             persistenceManager.numberOfActivations += 1
         }
+    #if os(visionOS)
+        .defaultSize(
+            width: 600,
+            height: 800
+        )
+    #endif
     }
     
     private func performLogging() {
         let systemVersion = UIDevice.current.systemVersion
         let version = Bundle.appVersion ?? "N/A"
         let build = Bundle.appBuild ?? "N/A"
-        let screenSize: CGRect = UIScreen.main.bounds
-        let screenWidth = screenSize.width.formatted()
-        let screenHeight = screenSize.height.formatted()
         let name = UIDevice.current.name
         
         logger.notice("OS Version: \(systemVersion, privacy: .public). App Version: \(version, privacy: .public) (\(build, privacy: .public)).")
-        logger.notice("Screen width: \(screenWidth, privacy: .public). Screen height: \(screenHeight, privacy: .public).")
         logger.notice("Device name: \(name, privacy: .public).")
+        
+        var screenSize: CGRect?
+        
+#if os(visionOS)
+        if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene {
+            screenSize = windowScene.coordinateSpace.bounds
+        }
+#else
+        screenSize = UIScreen.main.bounds
+#endif
+        
+        guard let screenSize else { return }
+        let screenWidth = screenSize.width.formatted()
+        let screenHeight = screenSize.height.formatted()
+        logger.notice("Screen width: \(screenWidth, privacy: .public). Screen height: \(screenHeight, privacy: .public).")
     }
 }
 
