@@ -10,7 +10,7 @@ import SwiftUI
 public struct SettingsMadeBy: View {
     let appID: Int
     @Environment(\.openURL) var openURL
-
+    
     public init(appID: Int){
         self.appID = appID
     }
@@ -18,6 +18,13 @@ public struct SettingsMadeBy: View {
     public var body: some View {
         VStack{
             VStack{
+#if os(visionOS)
+                Image("AppIcon_1024")
+                    .resizable()
+                    .scaledToFit()
+                    .frame(width: 35, height: 35)
+                    .cornerRadius(5)
+#else
                 if let icon = Bundle.appIcon(type: .current) {
                     Image(uiImage: icon)
                         .resizable()
@@ -25,13 +32,18 @@ public struct SettingsMadeBy: View {
                         .frame(width: 35, height: 35)
                         .cornerRadius(5)
                 }
+#endif
                 
                 Text(Bundle.appTitle ?? "N/A")
                     .font(.subheadline)
                     .fontWeight(.semibold)
                     .foregroundColor(Color(.systemGray))
                 Text("Version \(Bundle.appVersion ?? "N/A") (\(Bundle.appBuild ?? "N/A"))")
+#if os(visionOS)
+                    .foregroundColor(Color(.systemGray))
+#else
                     .foregroundColor(Color(.systemGray2))
+#endif
                     .font(.caption)
                     .padding(.bottom)
             }
@@ -43,7 +55,11 @@ public struct SettingsMadeBy: View {
                 Text("by ") + Text("Ricky Witherspoon")
                     .foregroundColor(.accentColor)
             }
+#if os(visionOS)
+            .foregroundColor(Color(.systemGray))
+#else
             .foregroundColor(Color(.systemGray2))
+#endif
             .font(.caption)
             .frame(maxWidth: .infinity, alignment: .center)
             .onTapGesture{
